fix: 级联删除如果父表没有删除的行，字表删除会失败。

+++ sharp-database/src/main/java/com/rick/db/plugin/dao/core/BaseDAOImpl.java
@@ -298,8 +298,10 @@ public class BaseDAOImpl<T, ID> implements BaseDAO<T, ID> {
         if (hasSubTables()) {
             List<ID> deletedIds = selectIdsByParams(params, conditionSQL);
 
-            for (String subTable : tableMeta.getSubTables()) {
-                SQLUtils.delete(subTable, subTableRefColumnName, deletedIds);
+            if (CollectionUtils.isNotEmpty(deletedIds)) {
+                for (String subTable : tableMeta.getSubTables()) {
+                    SQLUtils.delete(subTable, subTableRefColumnName, deletedIds);
+                }
             }
         }
